// src/app/models/tip.ts
export interface ITip {
  sourceid: number
  source: string
  year: number
  round: number
  gameid: number
  tipteamid: number
  tip: string
  date: string
  updated: string
  venue: string
  confidence: string
  hteamid: number
  hteam: string
  hconfidence: string
  ateamid: number
  ateam: string
  correct: number
  bits: string 
  err: string 
  margin: string
}